requirements.txt
fastapi==0.115.5
pydantic==2.10.2
SQLAlchemy==2.0.36
aiomysql==0.2.0
PyJWT~=2.10.1
passlib~=1.7.4
bcrypt==4.2.1
greenlet==3.1.1
typing_extensions~=4.12.2
cryptography~=44.0.0
pydantic-settings~=2.7.1